Extract @mentra/cloud-protocol so phones don't install the server runtime
The phone side (cloud-client's 10 importing files, engine's 3) consumed
@mentra/cloud-runtime/protocol exclusively — 10 files, 56K, zod-only —
while cloud-runtime's dependencies are the server's (ioredis,
@soniox/node, hono). Publishing runtime just to deliver its protocol
types would have shipped server internals to every phone consumer's
node_modules.

The protocol moves to its own leaf package (@mentra/cloud-protocol,
source-TS, zod only, wildcard subpath exports). Runtime's 16 internal
importers sweep to the package subpaths; its ./protocol entry stays as a
one-line compat shim; its tsconfig references the new project.
cloud-client and engine swap dependency and imports; engine's peer list
drops cloud-runtime entirely.

Consequence for the publish closure: cloud-runtime and cloud-shared
return to private (never published) — the set shrinks to jspolyfill,
crust, cloud-protocol, cloud-client, miniapp, and phone consumers trade
ioredis+soniox+hono+ulid for one small zod.

Verified: protocol 21 / runtime 34 / cloud-client 30 bun tests pass;
engine tsc + 220 tests; example OEM app typechecks; full mobile jest
55 suites / 429 tests; npm pack --dry-run clean for all five publish
packages; frozen installs pass on all three lockfiles.
