+++ .github/workflows/miniapp-sdk-release.yml
@@ -28,8 +28,7 @@ on:
       - "mobile/modules/miniapp/**"
       - "mobile/modules/jspolyfill/**"
       - "mobile/modules/crust/**"
-      - "cloud-v2/packages/shared/**"
-      - "cloud-v2/packages/runtime/**"
+      - "cloud-v2/packages/protocol/**"
       - "cloud-v2/packages/cloud-client/**"
       - "sdk/miniapp-cli/**"
       - "sdk/create-mentra-miniapp/**"
