+++ cloud-v2/packages/cloud-client/package.json
@@ -12,7 +12,7 @@
     "test": "bun test"
   },
   "dependencies": {
-    "@mentra/cloud-runtime": "^0.1.0-dev.0",
+    "@mentra/cloud-protocol": "^0.1.0-dev.0",
     "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
