+++ cloud-v2/packages/cloud-client/src/modules/runtime/runtime.reconnect.test.ts
@@ -34,7 +34,7 @@ import {
   type ConnectionInit,
   type ConnectionAck,
   type AudioSubscription,
-} from "@mentra/cloud-runtime/protocol";
+} from "@mentra/cloud-protocol";
 
 /** A scriptable WebSocket the test drives by hand (peer = the test). */
 class FakeSocket implements WebSocketLike {
