+++ cloud-v2/packages/protocol/src/camera.ts
