+++ cloud-v2/packages/protocol/src/errors.ts
