+++ cloud-v2/packages/protocol/src/handshake.ts
