+++ cloud-v2/packages/protocol/src/index.ts
@@ -1,0 +1,17 @@
+/**
+ * @fileoverview `@mentra/cloud-runtime/protocol` public entrypoint.
+ *
+ * Pure, isomorphic transport types + zod validators for the Mentra Runtime
+ * protocol. Zero server imports (no `node:*`, no service code) so the client
+ * can import the contract without pulling in server code.
+ *
+ * See docs/issues/002-cloud-runtime/protocol.md.
+ */
+export * from "./envelope";
+export * from "./handshake";
+export * from "./control";
+export * from "./errors";
+export * from "./messages";
+export * from "./audio";
+export * from "./camera";
+export * from "./maps";
