+++ cloud-v2/packages/protocol/tsconfig.json
@@ -5,6 +5,5 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "include": ["src/**/*"],
-  "references": [{ "path": "../shared" }, { "path": "../protocol" }]
+  "include": ["src/**/*"]
 }
