+++ cloud-v2/packages/runtime/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@mentra/cloud-runtime",
-  "version": "0.1.0-dev.0",
+  "version": "0.0.0",
+  "private": true,
   "type": "module",
   "main": "./src/index.ts",
   "exports": {
@@ -12,23 +13,13 @@
     "test": "bun test"
   },
   "dependencies": {
-    "@mentra/cloud-shared": "^0.1.0-dev.0",
+    "@mentra/cloud-protocol": "^0.1.0-dev.0",
+    "@mentra/cloud-shared": "workspace:*",
     "@soniox/node": "^2.1.0",
     "hono": "^4.11.1",
     "ioredis": "^5.4.1",
     "tweetnacl": "^1.0.3",
     "ulid": "^3.0.2",
     "zod": "^3.24.1"
-  },
-  "files": [
-    "src"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Mentra-Community/MentraOS.git",
-    "directory": "cloud-v2/packages/runtime"
   }
 }
