+++ cloud-v2/packages/runtime/src/services/session/subscriptions.ts
@@ -16,9 +16,9 @@ export type {
   TranscriptionSubscription,
   TranslationSubscription,
   AudioSubscription,
-} from "../../protocol/audio";
+} from "@mentra/cloud-protocol/audio";
 
-import type { AudioSubscription, LanguageSource } from "../../protocol/audio";
+import type { AudioSubscription, LanguageSource } from "@mentra/cloud-protocol/audio";
 
 /**
  * Stable string key for a subscription, used for equality + Map keys.
