+++ cloud-v2/packages/shared/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@mentra/cloud-shared",
-  "version": "0.1.0-dev.0",
+  "version": "0.0.0",
+  "private": true,
   "type": "module",
   "main": "./src/index.ts",
   "exports": {
@@ -19,16 +20,5 @@
   },
   "devDependencies": {
     "pino-pretty": "^11.3.0"
-  },
-  "files": [
-    "src"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Mentra-Community/MentraOS.git",
-    "directory": "cloud-v2/packages/shared"
   }
 }
